Fix negative VRAM values

+++ servers/rendering/renderer_rd/renderer_storage_rd.h
@@ -2263,7 +2263,7 @@ public:
 	void render_info_end_capture() {}
 	int get_captured_render_info(RS::RenderInfo p_info) { return 0; }
 
-	int get_render_info(RS::RenderInfo p_info) { return 0; }
+	uint64_t get_render_info(RS::RenderInfo p_info) { return 0; }
 	String get_video_adapter_name() const { return String(); }
 	String get_video_adapter_vendor() const { return String(); }
